requirements.txt
#Run pip install -r requirements.txt
#Requires python>=3.12
matplotlib
discord-py-interactions
apscheduler
pillow
python-dotenv